+++ Sources/Sentry/Public/SentryOptions.h
@@ -209,6 +209,16 @@ NS_SWIFT_NAME(Options)
 /**
  * This feature is EXPERIMENTAL.
  *
+ * Automatically attaches a textual representation of the view hierarchy when capturing an error
+ * event.
+ *
+ * Default value is <code>NO</code>
+ */
+@property (nonatomic, assign) BOOL attachViewHierarchy;
+
+/**
+ * This feature is EXPERIMENTAL.
+ *
  * When enabled, the SDK creates transactions for UI events like buttons clicks, switch toggles,
  * and other ui elements that uses UIControl `sendAction:to:forEvent:`.
  */