+++ Sources/Sentry/SentryBaseIntegration.m
@@ -134,6 +134,14 @@ NS_ASSUME_NONNULL_BEGIN
         return NO;
     }
 
+#if SENTRY_HAS_UIKIT
+    if ((integrationOptions & kIntegrationOptionAttachViewHierarchy)
+        && !options.attachViewHierarchy) {
+        [self logWithOptionName:@"attachViewHierarchy"];
+        return NO;
+    }
+#endif
+
     if ((integrationOptions & kIntegrationOptionEnableCrashHandler)
         && !options.enableCrashHandler) {
         [self logWithOptionName:@"enableCrashHandler"];