+++ Tests/SentryTests/Integrations/ViewHierarchy/TestSentryViewHierarchy.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+#if os(iOS) || os(tvOS) || targetEnvironment(macCatalyst)
+class TestSentryViewHierarchy: SentryViewHierarchy {
+
+    var result: [String] = []
+
+    override func fetch() -> [String] {
+        return result
+    }
+}
+#endif