+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Report App Memory Usage (#2027)
 - Include app permissions with event (#1984)
 - Add culture context to event (#2036)
+- Attach view hierarchy to events (#2044)
 - Clean up SentryOptions: added `enableCrashHandler` and deprecated `integrations` (#2049)
 
 ## 7.23.0